Add more functions documentaion and improve Makefile

+++ Makefile
@@ -1,7 +1,5 @@
-clear: 
-	rm -r Merge
-	rm -r Saida
-
-run: 
+run:
+	if [ -d "./Merge" ]; then rm -r Merge; fi
+	if [ -d "./Saida" ]; then rm -r Saida; fi
 	gcc -o teste *.c
 	valgrind ./teste
